rename more stages, move files
