+++ internal/handlers/common/aggregations/stages/pushdown.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aggregations
+package stages
 
 import (
 	"github.com/FerretDB/FerretDB/internal/types"
